refactor: organize things better
