+++ examples/sha2-chain/src/program.rs
@@ -1,0 +1,64 @@
+use zkaleido::{ProofType, ZkVmInputResult, ZkVmProgram, ZkVmProgramPerf};
+
+pub struct ShaChainProgram;
+
+impl ZkVmProgram for ShaChainProgram {
+    type Input = u32;
+    type Output = [u8; 32];
+
+    fn name() -> String {
+        "sha2_chain".to_string()
+    }
+
+    fn proof_type() -> zkaleido::ProofType {
+        ProofType::Core
+    }
+
+    fn prepare_input<'a, B>(input: &'a Self::Input) -> ZkVmInputResult<B::Input>
+    where
+        B: zkaleido::ZkVmInputBuilder<'a>,
+    {
+        B::new().write_serde(input)?.build()
+    }
+
+    fn process_output<H>(
+        public_values: &zkaleido::PublicValues,
+    ) -> zkaleido::ZkVmResult<Self::Output>
+    where
+        H: zkaleido::ZkVmHost,
+    {
+        H::extract_serde_public_output(public_values)
+    }
+}
+
+impl ZkVmProgramPerf for ShaChainProgram {}
+
+#[cfg(test)]
+mod tests {
+    use std::sync::Arc;
+
+    use zkaleido::ZkVmProgram;
+    use zkaleido_native_adapter::{NativeHost, NativeMachine};
+
+    use crate::{program::ShaChainProgram, process_sha2_chain};
+
+    fn get_native_host() -> NativeHost {
+        NativeHost {
+            process_proof: Arc::new(Box::new(move |zkvm: &NativeMachine| {
+                process_sha2_chain(zkvm);
+                Ok(())
+            })),
+        }
+    }
+
+    #[test]
+    fn test_native() {
+        let input = 5;
+        let host = get_native_host();
+        let receipt = ShaChainProgram::prove(&input, &host).unwrap();
+        let public_params =
+            ShaChainProgram::process_output::<NativeHost>(receipt.public_values()).unwrap();
+
+        assert!(public_params != [0; 32]);
+    }
+}
